move AgentId type
