+++ packages/server/src/services/agents/base/base-agent.ts
@@ -1,14 +1,14 @@
-import { z } from 'zod'
+import { extractEvents, extractTxWithEvents, flattenCalls, types } from '@sodazone/ocelloids-sdk'
 import { Operation } from 'rfc6902'
 import { Observable, from, share } from 'rxjs'
-import { extractEvents, extractTxWithEvents, flattenCalls, types } from '@sodazone/ocelloids-sdk'
+import { z } from 'zod'
 
-import { Logger, NetworkURN } from '../../types.js'
 import { IngressConsumer } from '../../ingress/index.js'
 import { NotifierHub } from '../../notification/hub.js'
 import { SubsStore } from '../../persistence/subs.js'
-import { AgentId, HexString, Subscription } from '../../subscriptions/types.js'
-import { Agent, AgentMetadata, AgentRuntimeContext } from '../types.js'
+import { HexString, Subscription } from '../../subscriptions/types.js'
+import { Logger, NetworkURN } from '../../types.js'
+import { Agent, AgentId, AgentMetadata, AgentRuntimeContext } from '../types.js'
 import { GetStorageAt } from '../xcm/types-augmented.js'
 
 type SubscriptionHandler = {
