+++ packages/server/src/services/subscriptions/api/ws/plugin.ts
@@ -1,7 +1,7 @@
 import { FastifyPluginAsync } from 'fastify'
 import fp from 'fastify-plugin'
 
-import { AgentId } from '../../types.js'
+import { AgentId } from '../../../agents/types.js'
 import WebsocketProtocol from './protocol.js'
 
 declare module 'fastify' {
