+++ packages/server/src/services/subscriptions/switchboard.ts
@@ -3,9 +3,9 @@ import EventEmitter from 'node:events'
 import { Operation } from 'rfc6902'
 
 import { Logger, Services } from '../types.js'
-import { AgentId, NotificationListener, Subscription, SubscriptionStats } from './types.js'
+import { NotificationListener, Subscription, SubscriptionStats } from './types.js'
 
-import { AgentService } from '../agents/types.js'
+import { AgentId, AgentService } from '../agents/types.js'
 import { NotifierEvents } from '../notification/types.js'
 import { TelemetryCollect, TelemetryEventEmitter } from '../telemetry/types.js'
